Render solution data from store
Note a letter + position pair may appear at most once
in the store, but other overlaps need to be
addressed when rendered. Fore example, we may learn that
a letter appears (but not in position 0) and then
later learn it appears in position 4.

Such a letter should not appear in the incorrect list.

Similarly, a letter which appears in correct or incorrect
should not appear in absent; e.g. PAPER with
first letter yellow; the third should not report as absent.