Add pseudo-absent letter checker
Close #5

This commit addresses the following problem with the solution
builder: Consider the guess PAPER for the correct answer
CUPID. Only the second P will be marked as green (correct),
which means that while P is certainly in the word (not absent)
it is also not the first letter (position 0), else that
would (also/instead) be green.

Such an absent letter is now categorized as pseudo-absent,
and this commit adds a checker for such an element.

As this letter has the absent flag, we need to also preserve
the position field in the data, so buildAbsentLetters
now takes an optional boolean argument (default false)
which, when true, keeps the position field and returns
an array of objects {letter, position} rather than
the only an array of letters.